refactor: move search_mixin to its own file
